Move getChildJsonConfig into #core/dom
